Refactor imports to be consistent
